+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## v2.0.3
 
+- Fix ComicInfo.xml coercing Age Rating to wrong value.
 - Normalize more original_format names.
 - Update xmltodict. Removes much comicbox fix code.